Major changes described bellow
Separate algorithms that take long time to process

Make two uwsgi caches so we never have to wait for
process to finish when we are asking for new data
