+++ tools/api/receiver.py
@@ -16,6 +16,7 @@ import pika
 import requests
 
 import tools.utility.log as log
+from tools.parseAndPopulate.modulesComplicatedAlgorithms import ModulesComplicatedAlgorithms
 from tools.utility import messageFactory
 from tools.utility.util import get_curr_dir
 
@@ -558,395 +559,11 @@ def on_request(ch, method, props, body):
             final_response = make_cache(credentials, final_response)
 
             if all_modules:
-                prefix = '{}://{}:{}'.format(confd_protocol, confd_ip,
-                                             confdPort)
-                new_modules = []
-                for module in all_modules['module']:
-                    LOGGER.info(
-                        'Searching semver for {}'.format(module['name']))
-                    url = '{}search/name/{}'.format(yangcatalog_api_prefix,
-                                                    module['name'])
-                    response = requests.get(url, auth=(
-                                            credentials[0], credentials[1]),
-                                            headers={
-                                                'Accept': 'application/json'})
-                    if response.status_code == 404:
-                        module['derived-semantic-version'] = '1.0.0'
-                        new_modules.append(module)
-                    else:
-                        data = json.loads(response.content)
-                        rev = module['revision'].split('-')
-                        date = datetime(int(rev[0]), int(rev[1]), int(rev[2]))
-                        module_temp = {}
-                        module_temp['name'] = module['name']
-                        module_temp['revision'] = module['revision']
-                        module_temp['organization'] = module['organization']
-                        module_temp['compilation'] = module[
-                            'compilation-status']
-                        module_temp['date'] = date
-                        module_temp['schema'] = module['schema']
-                        modules = [module_temp]
-                        semver_exist = True
-                        for mod in data['yang-catalog:modules']['module']:
-                            module_temp = {}
-                            revision = mod['revision']
-                            if revision == module['revision']:
-                                continue
-                            rev = revision.split('-')
-                            module_temp['revision'] = revision
-                            module_temp['date'] = datetime(int(rev[0]),
-                                                           int(rev[1]),
-                                                           int(rev[2]))
-                            module_temp['name'] = mod['name']
-                            module_temp['organization'] = mod['organization']
-                            module_temp['schema'] = mod.get('schema')
-                            module_temp['compilation'] = mod[
-                                'compilation-status']
-                            module_temp['semver'] = mod.get(
-                                'derived-semantic-version')
-                            if module_temp['semver'] is None:
-                                semver_exist = False
-                            modules.append(module_temp)
-
-                        if len(modules) == 1:
-                            module['derived-semantic-version'] = '1.0.0'
-                            new_modules.append(module)
-                            continue
-                        modules = sorted(modules, key=lambda k: k['date'])
-                        if modules[-1]['date'] == date and semver_exist:
-                            if modules[-1]['compilation'] != 'passed':
-                                versions = modules[-2]['semver'].split('.')
-                                ver = int(versions[0])
-                                ver += 1
-                                upgraded_version = '{}.{}.{}'.format(ver, 0, 0)
-                                module[
-                                    'derived-semantic-version'] = upgraded_version
-                                new_modules.append(module)
-                            else:
-                                if modules[-2]['compilation'] != 'passed':
-                                    versions = modules[-2]['semver'].split('.')
-                                    ver = int(versions[0])
-                                    ver += 1
-                                    upgraded_version = '{}.{}.{}'.format(ver, 0,
-                                                                         0)
-                                    module[
-                                        'derived-semantic-version'] = upgraded_version
-                                    new_modules.append(module)
-                                    continue
-                                else:
-                                    schema2 = '{}{}@{}.yang'.format(
-                                        save_file_dir,
-                                        modules[-2]['name'],
-                                        modules[-2]['revision'])
-                                    schema1 = '{}{}@{}.yang'.format(
-                                        save_file_dir,
-                                        modules[-1]['name'],
-                                        modules[-1]['revision'])
-                                arguments = ['pyang', '-P', get_curr_dir(__file__) + '/../../.', '-p',
-                                             get_curr_dir(__file__) + '/../../.',
-                                             schema1, '--check-update-from',
-                                             schema2]
-                                pyang = subprocess.Popen(arguments,
-                                                         stdout=subprocess.PIPE,
-                                                         stderr=subprocess.PIPE)
-                                stdout, stderr = pyang.communicate()
-                                if stderr == '':
-                                    arguments = ["pyang", '-p', get_curr_dir(__file__) + '/../../.', "-f",
-                                                 "tree",
-                                                 schema1]
-                                    pyang = subprocess.Popen(arguments,
-                                                             stdout=subprocess.PIPE,
-                                                             stderr=subprocess.PIPE)
-                                    stdout, stderr = pyang.communicate()
-                                    arguments = ["pyang", "-p", get_curr_dir(__file__) + "/../../.", "-f",
-                                                 "tree",
-                                                 schema2]
-                                    pyang = subprocess.Popen(arguments,
-                                                             stdout=subprocess.PIPE,
-                                                             stderr=subprocess.PIPE)
-                                    stdout2, stderr = pyang.communicate()
-                                    if stdout == stdout2:
-                                        versions = modules[-2]['semver'].split(
-                                            '.')
-                                        ver = int(versions[2])
-                                        ver += 1
-                                        upgraded_version = '{}.{}.{}'.format(
-                                            versions[0],
-                                            versions[1],
-                                            ver)
-                                        module[
-                                            'derived-semantic-version'] = upgraded_version
-                                        new_modules.append(module)
-                                        continue
-                                    else:
-                                        versions = modules[-2]['semver'].split(
-                                            '.')
-                                        ver = int(versions[1])
-                                        ver += 1
-                                        upgraded_version = '{}.{}.{}'.format(
-                                            versions[0],
-                                            ver, 0)
-                                        module[
-                                            'derived-semantic-version'] = upgraded_version
-                                        new_modules.append(module)
-                                        continue
-                                else:
-                                    versions = modules[-2]['semver'].split('.')
-                                    ver = int(versions[0])
-                                    ver += 1
-                                    upgraded_version = '{}.{}.{}'.format(ver, 0,
-                                                                         0)
-                                    module[
-                                        'derived-semantic-version'] = upgraded_version
-                                    new_modules.append(module)
-                                    continue
-                        else:
-                            mod = {}
-                            mod['name'] = modules[0]['name']
-                            mod['revision'] = modules[0]['revision']
-                            mod['organization'] = modules[0]['organization']
-                            modules[0]['semver'] = '1.0.0'
-                            response = requests.get(
-                                '{}://{}:{}/api/config/catalog/modules/module/{},{},{}'.format(
-                                    protocol, confd_ip, confdPort,
-                                    mod['name'], mod['revision'],
-                                    mod['organization']),
-                                auth=(credentials[0], credentials[1]), headers={
-                                    'Accept': 'application/vnd.yang.data+json'})
-                            response = json.loads(response.content)[
-                                'yang-catalog:module']
-                            response['derived-semantic-version'] = '1.0.0'
-                            new_modules.append(response)
-
-                            for x in range(1, len(modules)):
-                                mod = {}
-                                mod['name'] = modules[x]['name']
-                                mod['revision'] = modules[x]['revision']
-                                mod['organization'] = modules[x]['organization']
-                                if modules[x]['compilation'] != 'passed':
-                                    versions = modules[x - 1]['semver'].split(
-                                        '.')
-                                    ver = int(versions[0])
-                                    ver += 1
-                                    upgraded_version = '{}.{}.{}'.format(ver, 0,
-                                                                         0)
-                                    modules[x]['semver'] = upgraded_version
-                                    response = requests.get(
-                                        '{}://{}:{}/api/config/catalog/modules/module/{},{},{}'.format(
-                                            protocol, confd_ip, confdPort,
-                                            mod['name'], mod['revision'],
-                                            mod['organization']),
-                                        auth=(credentials[0], credentials[1]), headers={
-                                            'Accept': 'application/vnd.yang.data+json'})
-                                    response = json.loads(response.content)[
-                                        'yang-catalog:module']
-                                    response[
-                                        'derived-semantic-version'] = upgraded_version
-                                    new_modules.append(response)
-                                else:
-                                    if modules[x - 1][
-                                        'compilation'] != 'passed':
-                                        versions = modules[x - 1][
-                                            'semver'].split('.')
-                                        ver = int(versions[0])
-                                        ver += 1
-                                        upgraded_version = '{}.{}.{}'.format(
-                                            ver, 0, 0)
-                                        modules[x]['semver'] = upgraded_version
-                                        response = requests.get(
-                                            '{}://{}:{}/api/config/catalog/modules/module/{},{},{}'.format(
-                                                protocol, confd_ip, confdPort,
-                                                mod['name'], mod['revision'],
-                                                mod['organization']),
-                                            auth=(
-                                            credentials[0], credentials[1]), headers={
-                                                'Accept': 'application/vnd.yang.data+json'})
-                                        response = json.loads(response.content)[
-                                            'yang-catalog:module']
-                                        response[
-                                            'derived-semantic-version'] = upgraded_version
-                                        new_modules.append(response)
-                                        continue
-                                    else:
-                                        schema2 = '{}{}@{}.yang'.format(
-                                            save_file_dir,
-                                            modules[x]['name'],
-                                            modules[x]['revision'])
-                                        schema1 = '{}{}@{}.yang'.format(
-                                            save_file_dir,
-                                            modules[x - 1]['name'],
-                                            modules[x - 1]['revision'])
-                                    arguments = ['pyang', '-p', get_curr_dir(__file__) + '/../../.', '-P',
-                                                 get_curr_dir(__file__) + '/../../.',
-                                                 schema2,
-                                                 '--check-update-from', schema1]
-                                    pyang = subprocess.Popen(arguments,
-                                                             stdout=subprocess.PIPE,
-                                                             stderr=subprocess.PIPE)
-                                    stdout, stderr = pyang.communicate()
-                                    if stderr == '':
-                                        arguments = ["pyang", '-p', get_curr_dir(__file__) + '/../../.',
-                                                     "-f", "tree",
-                                                     schema1]
-                                        pyang = subprocess.Popen(arguments,
-                                                                 stdout=subprocess.PIPE,
-                                                                 stderr=subprocess.PIPE)
-                                        stdout, stderr = pyang.communicate()
-                                        arguments = ["pyang", '-p', get_curr_dir(__file__) + '/../../.',
-                                                     "-f", "tree",
-                                                     schema2]
-                                        pyang = subprocess.Popen(arguments,
-                                                                 stdout=subprocess.PIPE,
-                                                                 stderr=subprocess.PIPE)
-                                        stdout2, stderr = pyang.communicate()
-                                        if stdout == stdout2:
-                                            versions = modules[x - 1][
-                                                'semver'].split('.')
-                                            ver = int(versions[2])
-                                            ver += 1
-                                            upgraded_version = '{}.{}.{}'.format(
-                                                versions[0], versions[1], ver)
-                                            modules[x][
-                                                'semver'] = upgraded_version
-                                            response = requests.get(
-                                                '{}://{}:{}/api/config/catalog/modules/module/{},{},{}'.format(
-                                                    protocol, confd_ip,
-                                                    confdPort,
-                                                    mod['name'],
-                                                    mod['revision'],
-                                                    mod['organization']),
-                                                auth=(
-                                                credentials[0], credentials[1]),
-                                                headers={
-                                                    'Accept': 'application/vnd.yang.data+json'})
-                                            response = \
-                                            json.loads(response.content)[
-                                                'yang-catalog:module']
-                                            response[
-                                                'derived-semantic-version'] = upgraded_version
-                                            new_modules.append(response)
-                                        else:
-                                            versions = modules[x - 1][
-                                                'semver'].split('.')
-                                            ver = int(versions[1])
-                                            ver += 1
-                                            upgraded_version = '{}.{}.{}'.format(
-                                                versions[0], ver, 0)
-                                            modules[x][
-                                                'semver'] = upgraded_version
-                                            response = requests.get(
-                                                '{}://{}:{}/api/config/catalog/modules/module/{},{},{}'.format(
-                                                    protocol, confd_ip,
-                                                    confdPort,
-                                                    mod['name'],
-                                                    mod['revision'],
-                                                    mod['organization']),
-                                                auth=(
-                                                credentials[0], credentials[1]),
-                                                headers={
-                                                    'Accept': 'application/vnd.yang.data+json'})
-                                            response = \
-                                            json.loads(response.content)[
-                                                'yang-catalog:module']
-                                            response[
-                                                'derived-semantic-version'] = upgraded_version
-                                            new_modules.append(response)
-                                    else:
-                                        versions = modules[x - 1][
-                                            'semver'].split('.')
-                                        ver = int(versions[0])
-                                        ver += 1
-                                        upgraded_version = '{}.{}.{}'.format(
-                                            ver, 0, 0)
-                                        modules[x]['semver'] = upgraded_version
-                                        response = requests.get(
-                                            '{}://{}:{}/api/config/catalog/modules/module/{},{},{}'.format(
-                                                protocol, confd_ip,
-                                                confdPort,
-                                                mod['name'], mod['revision'],
-                                                mod['organization']),
-                                            auth=(
-                                            credentials[0], credentials[1]), headers={
-                                                'Accept': 'application/vnd.yang.data+json'})
-                                        response = json.loads(response.content)[
-                                            'yang-catalog:module']
-                                        response[
-                                            'derived-semantic-version'] = upgraded_version
-                                        new_modules.append(response)
-
-                for mod in all_modules['module']:
-                    name = mod['name']
-                    revision = mod['revision']
-                    new_dependencies = mod['dependencies']
-                    for new_dep in new_dependencies:
-                        if new_dep.get('revision'):
-                            search = {'name': new_dep['name'],
-                                      'revision': new_dep['revision']}
-                        else:
-                            search = {'name': new_dep['name']}
-                        response = requests.post(yangcatalog_api_prefix
-                                                 + 'search-filter',
-                            auth=(credentials[0], credentials[1]),
-                            json={'input': search})
-                        if response.status_code == 200:
-                            mods = \
-                            json.loads(response.content)['yang-catalog:modules'][
-                                'module']
-                            for m in mods:
-                                if m.get('dependents') is None:
-                                    m['dependents'] = []
-                                new = {'name': name,
-                                       'revision': revision,
-                                       'schema': mod['schema']}
-                                if new not in m['dependents']:
-                                    m['dependents'].append(new)
-                                    new_modules.append(m)
-
-                    response = requests.post(yangcatalog_api_prefix
-                                             + 'search-filter',
-                        auth=(
-                            credentials[0], credentials[1]),
-                        json={'input': {'dependencies': [{'name': name}]}})
-                    if response.status_code == 200:
-                        mods = json.loads(response.content)['yang-catalog:modules'][
-                            'module']
-
-                        if mod.get('dependents') is None:
-                            mod['dependents'] = []
-                        for m in mods:
-                            passed = False
-                            for dependency in m['dependencies']:
-                                if dependency['name'] == name:
-                                    passed = True
-                                    rev = dependency.get('revision')
-                                    if rev:
-                                        passed = False
-                                        if rev == revision:
-                                            passed = True
-                            if passed:
-                                new = {'name': m['name'],
-                                       'revision': m['revision'],
-                                       'schema': m['schema']}
-                                if new not in mod['dependents']:
-                                    mod['dependents'].append(new)
-                        if len(mod['dependents']) > 0:
-                            new_modules.append(mod)
-                mod = len(new_modules) % 1000
-                for x in range(0, len(new_modules) / 1000):
-                    json_modules_data = json.dumps({'modules': {
-                        'module': new_modules[x * 1000: (x * 1000) + 1000]}})
-                    if '{"module": []}' not in json_modules_data:
-                        http_request(prefix + '/api/config/catalog/modules/',
-                                     'PATCH',
-                                     json_modules_data, credentials,
-                                     'application/vnd.yang.data+json')
-                rest = (len(new_modules) / 1000) * 1000
-                json_modules_data = json.dumps(
-                    {'modules': {'module': new_modules[rest: rest + mod]}})
-                if '{"module": []}' not in json_modules_data:
-                    http_request(prefix + '/api/config/catalog/modules/', 'PATCH',
-                                 json_modules_data, credentials,
-                                 'application/vnd.yang.data+json')
+                complicatedAlgorithms = ModulesComplicatedAlgorithms(yangcatalog_api_prefix, credentials,
+                                                                     confd_protocol, confd_ip, confdPort,
+                                                                     save_file_dir, None, all_modules)
+                complicatedAlgorithms.parse(False)
+                complicatedAlgorithms.populate()
 
     LOGGER.info('Receiver is done with id - {} and message = {}'
                 .format(props.correlation_id, str(final_response)))
