+++ tools/parseAndPopulate/modules.py
@@ -223,7 +223,6 @@ class Modules:
             self.__resolve_author_email(author_email)
             self.__resolve_maturity_level(maturity_level)
             self.__resolve_semver()
-            self.__resolve_tree_type()
 
     def __resolve_tree(self):
         if self.module_type == 'module':
@@ -354,282 +353,6 @@ class Modules:
                 self.schema = schema
         else:
             self.schema = None
-
-    def __resolve_tree_type(self):
-        def is_openconfig(rows, output):
-            count_config = output.count('+-- config')
-            count_state = output.count('+-- state')
-            if count_config != count_state:
-                return False
-            row_number = 0
-            skip = []
-            for row in rows:
-                if 'x--' in row or 'o--' in row:
-                    continue
-                if '' == row.strip(' '):
-                    break
-                if '+--rw' in row and row_number != 0 \
-                        and row_number not in skip and '[' not in row and \
-                        (len(row.replace('|', '').strip(' ').split(
-                            ' ')) != 2 or '(' in row):
-                    if '->' in row and 'config' in row.split('->')[
-                        1] and '+--rw config' not in rows[row_number - 1]:
-                        row_number += 1
-                        continue
-                    if '+--rw config' not in rows[row_number - 1]:
-                        if 'augment' in rows[row_number - 1]:
-                            if not rows[row_number - 1].endswith(':config:'):
-                                return False
-                        else:
-                            return False
-                    length_before = set([len(row.split('+--')[0])])
-                    skip = []
-                    for x in range(row_number, len(rows)):
-                        if 'x--' in rows[x] or 'o--' in rows[x]:
-                            continue
-                        if len(rows[x].split('+--')[0]) not in length_before:
-                            if (len(rows[x].replace('|', '').strip(' ').split(
-                                    ' ')) != 2 and '[' not in rows[x]) \
-                                    or '+--:' in rows[x] or '(' in rows[x]:
-                                length_before.add(len(rows[x].split('+--')[0]))
-                            else:
-                                break
-                        if '+--ro' in rows[x]:
-                            return False
-                        duplicate = \
-                            rows[x].replace('+--rw', '+--ro').split('+--')[1]
-                        if duplicate.replace(' ', '') not in output.replace(' ',
-                                                                            ''):
-                            return False
-                        skip.append(x)
-                if '+--ro' in row and row_number != 0 and row_number not in skip and '[' not in row and \
-                        (len(row.replace('|', '').strip(' ').split(
-                            ' ')) != 2 or '(' in row):
-                    if '->' in row and 'state' in row.split('->')[
-                        1] and '+--ro state' not in rows[row_number - 1]:
-                        row_number += 1
-                        continue
-                    if '+--ro state' not in rows[row_number - 1]:
-                        if 'augment' in rows[row_number - 1]:
-                            if not rows[row_number - 1].endswith(':state:'):
-                                return False
-                        else:
-                            return False
-                    length_before = len(row.split('+--')[0])
-                    skip = []
-                    for x in range(row_number, len(rows)):
-                        if 'x--' in rows[x] or 'o--' in rows[x]:
-                            continue
-                        if len(rows[x].split('+--')[0]) < length_before:
-                            break
-                        if '+--rw' in rows[x]:
-                            return False
-                        skip.append(x)
-                row_number += 1
-            return True
-
-        def is_combined(rows, output):
-            if output.split('\n')[0].endswith('-state'):
-                return False
-            next_obsolete_or_deprecated = False
-            for row in rows:
-                if next_obsolete_or_deprecated:
-                    if 'x--' in row or 'o--' in row:
-                        next_obsolete_or_deprecated = False
-                    else:
-                        return False
-                if 'x--' in row or 'o--' in row:
-                    continue
-                if '+--rw config' == row.replace('|', '').strip(
-                        ' ') or '+--ro state' == row.replace('|', '').strip(
-                    ' '):
-                    return False
-                if len(row.split('+--')[0]) == 4:
-                    if '-state' in row and '+--ro' in row:
-                        return False
-                if len(row.split('augment')[0]) == 2:
-                    part = row.strip(' ').split('/')[1]
-                    if '-state' in part:
-                        next_obsolete_or_deprecated = True
-                    part = row.strip(' ').split('/')[-1]
-                    if ':state:' in part or '/state:' in part \
-                            or ':config:' in part or '/config:' in part:
-                        next_obsolete_or_deprecated = True
-            return True
-
-        def is_transational(rows, output):
-            if output.split('\n')[0].endswith('-state'):
-                if '+--rw' in output:
-                    return False
-                name_of_module = output.split('\n')[0].split(': ')[1]
-                name_of_module = name_of_module.split('-state')[0]
-                coresponding_nmda_file = self.__find_file(name_of_module,
-                                                          normal_search=False)
-                if coresponding_nmda_file:
-                    arguments = ["pyang", "-p", get_curr_dir(__file__) + "/../../.", "-f", "tree",
-                                 coresponding_nmda_file]
-                    pyang = subprocess.Popen(arguments, stdout=PIPE,
-                                             stderr=PIPE)
-                    stdout, stderr = pyang.communicate()
-                    pyang_list_of_rows = stdout.split('\n')[1:]
-                    if 'error' in stderr and 'is not found' in stderr:
-                        return False
-                    elif stdout == '':
-                        return False
-                    for x in range(0, len(rows)):
-                        if 'x--' in rows[x] or 'o--' in rows[x]:
-                            continue
-                        if rows[x].strip(' ') == '':
-                            break
-                        if len(rows[x].split('+--')[0]) == 4:
-                            if '-state' in rows[x]:
-                                return False
-                        if len(rows[x].split('augment')[0]) == 2:
-                            part = rows[x].strip(' ').split('/')[1]
-                            if '-state' in part:
-                                return False
-                        if '+--ro ' in rows[x]:
-                            leaf = \
-                                rows[x].split('+--ro ')[1].split(' ')[0].split(
-                                    '?')[0]
-
-                            dataExist = False
-                            for y in range(0, len(pyang_list_of_rows)):
-                                if leaf in pyang_list_of_rows[y]:
-                                    dataExist = True
-                            if not dataExist:
-                                return False
-                    return True
-                else:
-                    return False
-            else:
-                return False
-
-        def is_split(rows, output):
-            #states = set()
-            failed = False
-            row_num = 0
-            if output.split('\n')[0].endswith('-state'):
-                return False
-            for row in rows:
-                if 'x--' in row or 'o--' in row:
-                    continue
-                if '+--rw config' == row.replace('|', '').strip(
-                        ' ') or '+--ro state' == row.replace('|', '')\
-                        .strip(' '):
-                    return False
-                if 'augment' in row:
-                    part = row.strip(' ').split('/')[-1]
-                    if ':state:' in part or '/state:' in part or ':config:' in part or '/config:' in part:
-                        return False
-            for row in rows:
-                if 'x--' in row or 'o--' in row:
-                    continue
-                if row == '':
-                    break
-                if (len(row.split('+--')[0]) == 4 and 'augment' not in rows[
-                        row_num - 1]) or len(row.split('augment')[0]) == 2:
-                    if '-state' in row:
-                        if 'augment' in row:
-                            part = row.strip(' ').split('/')[1]
-                            if '-state' not in part:
-                                row_num += 1
-                                continue
-                            #if ':' in part:
-                            #    state = part.split(':')[1].split('-state')[0]
-                            #else:
-                            #    state = part.split('-state')[0]
-                        #else:
-                        #    state = \
-                        #        row.strip(' ').split('-state')[0].split(' ')[1]
-                        for x in range(row_num + 1, len(rows)):
-                            if 'x--' in rows[x] or 'o--' in rows[x]:
-                                continue
-                            if rows[x].strip(' ') == '' \
-                                    or (len(rows[x].split('+--')[
-                                                0]) == 4 and 'augment' not in
-                                        rows[row_num - 1]) \
-                                    or len(row.split('augment')[0]) == 2:
-                                break
-                            if '+--rw' in rows[x]:
-                                failed = True
-                                break
-                        #states.add(state)
-
-                row_num += 1
-            #if failed:
-            #    return False
-#
-            #for state in states:
-            #    if failed:
-            #        return False
-            #    row_num = 0
-            #    failed = True
-            #    for row in rows:
-            #        if 'x--' in row or 'o--' in row:
-            #            continue
-            #        if '+--:' in row or '+--ro' in row:
-            #            continue
-            #        if row.strip(' ') == '':
-            #            failed = True
-            #            break
-            #        if (len(row.split('+--')[0]) == 4 and 'augment' not in rows[
-            #                row_num - 1]) \
-            #                or len(row.split('augment')[0]) == 2:
-            #            if ('augment' in row and (
-            #                                    ':' + state + '/' in row or '/' + state + '/' in row)) \
-            #                    or ('augment' in row
-            #                        and (
-            #                                                    ':' + state + '-config' + '/' in row or '/' + state + '-config' + '/' in row)) \
-            #                    or (state + '-config' ==
-            #                            row.strip(' ').split(' ')[
-            #                                1]) \
-            #                    or (state == row.strip(' ').split(' ')[1]):
-            #                failed = False
-            #                for x in range(row_num + 1, len(rows)):
-            #                    if 'x--' in rows[x] or 'o--' in rows[x]:
-            #                        continue
-            #                    if rows[x].strip(' ') == '' or len(
-            #                            rows[x].split('+--')[0]) == 4 \
-            #                            or len(row.split('augment')[0]) == 2:
-            #                        break
-            #                    if '+--ro' in rows[x]:
-            #                        failed = True
-            #                        break
-            #                break
-            #        row_num += 1
-            if failed:
-                return False
-            else:
-                return True
-
-        LOGGER.debug(
-            'Get tree type from tag from module {}'.format(self.__path))
-        arguments = ["pyang", "-p", get_curr_dir(__file__) + "/../../.", "-f", "tree", self.__path]
-        pyang = subprocess.Popen(arguments, stdout=PIPE, stderr=PIPE)
-        stdout, stderr = pyang.communicate()
-        if 'error' in stderr and 'is not found' in stderr:
-            LOGGER.debug(
-                'Could not use pyang to generate tree because of error {} on module {}'.
-                    format(stderr, self.__path))
-            self.tree_type = 'unclassified'
-        elif stdout == '':
-            self.tree_type = 'not-applicable'
-        else:
-            pyang_list_of_rows = stdout.split('\n')[1:]
-            if 'submodule' == self.module_type:
-                LOGGER.debug('Module {} is a submodule'.format(self.__path))
-                self.tree_type = 'not-applicable'
-            elif is_combined(pyang_list_of_rows, stdout):
-                self.tree_type = 'nmda-compatible'
-            elif is_transational(pyang_list_of_rows, stdout):
-                self.tree_type = 'transitional-extra'
-            elif is_openconfig(pyang_list_of_rows, stdout):
-                self.tree_type = 'openconfig'
-            elif is_split(pyang_list_of_rows, stdout):
-                self.tree_type = 'split'
-            else:
-                self.tree_type = 'unclassified'
 
     def __resolve_module_classification(self, module_classification=None):
         if module_classification:
